# api/docs/swagger.yaml
definitions:
...
  model_common.ResponseError:
    properties:
      data:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  model_common.StandardErrorModel:
    properties:
      error:
        $ref: '#/definitions/model_common.ResponseError'
    type: object
...
  model_healthcare_service.DoctorAndDoctorHours:
    properties:
      address:
        type: string
      bio:
        type: string
      birth_date:
        type: string
      city:
        type: string
      country:
        type: string
      created_at:
        type: string
      day_of_week:
        type: string
      deleted_at:
        type: string
      department_id:
        type: string
      email:
        type: string
      end_work_date:
        type: string
      finish_time:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: string
      image_url:
        type: string
      last_name:
        type: string
      order:
        type: integer
      patient_count:
        type: integer
      phone_number:
        type: string
      room_number:
        type: integer
      salary:
        type: number
      specializations:
        items:
          $ref: '#/definitions/model_healthcare_service.DoctorSpec'
        type: array
      start_time:
        type: string
      start_work_date:
        type: string
      updated_at:
        type: string
      work_years:
        type: integer
    type: object
  model_healthcare_service.DoctorReq:
    properties:
      address:
        example: Addres
        type: string
      bio:
        example: Biography
        type: string
      birth_date:
        example: "2012-12-12"
        type: string
      city:
        example: City
        type: string
      country:
        example: Country
        type: string
      department_id:
        example: 123e4567-e89b-12d3-a456-426614174001
        type: string
      email:
        example: email@gmail.com
        type: string
      first_name:
        example: First Name
        type: string
      gender:
        example: male
        type: string
      image_url:
        example: http://example.com/image.png
        type: string
      last_name:
        example: Last Name
        type: string
      password:
        example: password
        type: string
      phone_number:
        example: "+998901234567"
        type: string
      room_number:
        example: 1
        type: integer
      salary:
        example: 10
        type: number
      start_work_date:
        example: "2012-12-12"
        type: string
      work_years:
        example: 4
        type: integer
    type: object
  model_healthcare_service.DoctorRes:
    properties:
      address:
        type: string
      bio:
        type: string
      birth_date:
        type: string
      city:
        type: string
      country:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      department_id:
        type: string
      email:
        type: string
      end_work_date:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: string
      image_url:
        type: string
      last_name:
        type: string
      order:
        type: integer
      phone_number:
        type: string
      room_number:
        type: integer
      salary:
        type: number
      start_work_date:
        type: string
      updated_at:
        type: string
      work_years:
        type: integer
    type: object
...
  model_healthcare_service.DoctorSpec:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  model_healthcare_service.DoctorUpdateReq:
    properties:
      address:
        example: Addres
        type: string
      bio:
        example: Biography
        type: string
      birth_date:
        example: "2012-12-12"
        type: string
      city:
        example: City
        type: string
      country:
        example: Country
        type: string
      department_id:
        example: 123e4567-e89b-12d3-a456-426614174001
        type: string
      email:
        example: email@gmail.com
        type: string
      end-work-date:
        example: "2022-12-12"
        type: string
      first_name:
        example: First Name
        type: string
      gender:
        example: male
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614274001
        type: string
      image_url:
        example: http://example.com/image.png
        type: string
      last_name:
        example: Last Name
        type: string
      password:
        example: password
        type: string
      phone_number:
        example: "+998901234567"
        type: string
      room_number:
        example: 1
        type: integer
      salary:
        example: 10
        type: number
      start_work_date:
        example: "2012-12-12"
        type: string
      work_years:
        example: 4
        type: integer
    type: object
...
  model_healthcare_service.ListDoctors:
    properties:
      count:
        type: integer
      doctors:
        items:
          $ref: '#/definitions/model_healthcare_service.DoctorRes'
        type: array
    type: object
...
  models.StatusRes:
    properties:
      status:
        type: boolean
    type: object
host: localhost:9050
info:
  contact: {}
  title: Dennic Project
  version: "1.7"
paths:
...
  /v1/doctor:
    delete:
      consumes:
      - application/json
      description: DeleteDoctor - Api for delete doctor
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StatusRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
      summary: DeleteDoctor
      tags:
      - Doctor
    get:
      consumes:
      - application/json
      description: ListDoctors - Api for list doctor
      parameters:
      - in: query
        name: limit
        type: string
      - in: query
        name: order_by
        type: string
      - in: query
        name: page
        type: string
      - in: query
        name: value
        type: string
      - description: search
        enum:
        - first_name
        - last_name
        - gender
        - phone_number
        - email
        - address
        - city
        - country
        - biography
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model_healthcare_service.DoctorAndDoctorHours'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
      summary: ListDoctors
      tags:
      - Doctor
    post:
      consumes:
      - application/json
      description: CreateDoctor - Api for crete doctor
      parameters:
      - description: DoctorReq
        in: body
        name: DoctorReq
        required: true
        schema:
          $ref: '#/definitions/model_healthcare_service.DoctorReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model_healthcare_service.DoctorRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
      summary: CreateDoctor
      tags:
      - Doctor
    put:
      consumes:
      - application/json
      description: UpdateDoctor - Api for update doctor
      parameters:
      - description: UpdateDoctorReq
        in: body
        name: UpdateDoctorReq
        required: true
        schema:
          $ref: '#/definitions/model_healthcare_service.DoctorUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model_healthcare_service.DoctorRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
      summary: UpdateDoctor
      tags:
      - Doctor
...
  /v1/doctor/get:
    get:
      consumes:
      - application/json
      description: GetDoctor - Api for get doctor
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model_healthcare_service.DoctorAndDoctorHours'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
      summary: GetDoctor
      tags:
      - Doctor
  /v1/doctor/spec:
    get:
      consumes:
      - application/json
      description: ListDoctorsBySpecializationId - Api for list doctors by specialization
        id
      parameters:
      - in: query
        name: limit
        type: string
      - in: query
        name: order_by
        type: string
      - in: query
        name: page
        type: string
      - in: query
        name: value
        type: string
      - description: specialization_id
        in: query
        name: specialization_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model_healthcare_service.ListDoctors'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model_common.StandardErrorModel'
      summary: ListDoctorsBySpecializationId
      tags:
      - Doctor
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"